requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
altair
OpenAI
tiktoken
streamlit-extras
pinecone-client
mysql-connector-python